requirements.txt
black==24.10.0
crc==7.1.0
isort==5.13.2
Pillow==11.0.0